pyproject.toml
[tool.poetry]
name = "daxiang_trade"
version = "0.1.0"
description = ""
authors = ["Tang Wei <tw7613781@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
websocket-client = "^1.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
